fix: separate the downloading module
